Adding PermissionStats.java
Adding Java program to analyze the APKs extracted by checking their AndroidManifest.xml file
