+++ .gitignore
@@ -1,3 +1,6 @@
+*.xml
+*.iml
+
 # Compiled class file
 *.class
